requirements.txt
numpy 
pandas 
python-dotenv
pymongo
certifi
pymongo[srv]==3.10
scikit-learn
scipy
dill
pickle
pyaml
#-e . 
